Speeding_up_boot

+++ deploy.py
@@ -1,7 +1,7 @@
 import glob
 import shutil
 import os
-from pathlib import Path
+import filecmp
 
 files = [
     "code.py"
@@ -20,11 +20,14 @@ destination = os.path.abspath(destination)
 def deploy(file_name):
     source = os.path.join(path, file_name)
     final_destination = os.path.join(destination, file_name)
-    # os.makedirs(final_destination, exist_ok=True)
+
     file_directory = os.path.dirname(final_destination)
     if(not os.path.exists(file_directory)):
         print("Making directory: {}".format(file_directory))
         os.makedirs(file_directory, exist_ok=True)
+    if(os.path.exists(final_destination) and filecmp.cmp(source, final_destination, shallow=False)):
+        print("no change {}".format(final_destination))
+        return
     print("Copying {} to {}".format(source, final_destination))
     shutil.copy(source, final_destination)
